pyproject.toml
[tool.poetry]
name = "jiggymusicbot"
version = "0.1.0"
description = "A Discord music bot"
authors = ["Jiggy Jacobs"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
discord-py = "2.4.0"
youtube-dl = "2021.12.17"
ffmpeg-python = "0.2.0"
python-dotenv = "^1.0.1"
yt-dlp = "^2024.8.6"
flask = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
